Shuffle jobs in worker pipeline
Move more work to the url fetching component of the worker thread
pipeline and improve error handling. The slight delay in the
spawning of the next thread should even out the workload a bit
more to increase throughput. In addition, improve the error
handling of errors relevent to replica sets.

Signed-off-by: Daniel Robertson <danlrobertson89@gmail.com>