Fix unit test ImportError for palantiri
The previous method of running unit tests involved a subprocess on
calling `python -m unittest` which could error on a system with
multiple Python installs. The specific error has been documented in #8.

ImportError was solved by removing the `cmd_class` re-routing the
`python setup.py test` command. setuptools was used to in place of
distutils to support `install_requires` functionality.